config/urls.py
from django.conf.urls import url, include
from django.conf.urls.static import static

from django.contrib import admin
from django.views.generic import RedirectView
from django.conf import settings

urlpatterns = [
    url(r'^admin/', admin.site.urls),
    url(r'^$', RedirectView.as_view(url='/accounts/login/')),
    url(r'^accounts/', include('accounts.urls')),
    url(r'^shop/', include('shop.urls')),
]

if settings.DEBUG:
    import debug_toolbar

    urlpatterns += [
        url(r'^__debug__/', include(debug_toolbar.urls)),
    ]

# TODO: https://docs.djangoproject.com/ja/1.11/howto/static-files/#serving-files-uploaded-by-a-user-during-development
if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
